+++ src/main/java/com/example/model/Agent.java
@@ -5,5 +5,5 @@ import javafx.beans.property.BooleanProperty;
 public interface Agent {
     BooleanProperty isTraitor();
     void setIsTraitor(boolean isTraitor);
-    IOpinions getOpinions();
+    Opinions getOpinions();
 }
