+++ src/main/java/com/example/model/AgentOpinion.java
@@ -1,0 +1,18 @@
+package com.example.model;
+
+import javafx.beans.property.BooleanProperty;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+public class AgentOpinion implements Opinion {
+    @Getter
+    private final String name;
+    @Getter
+    private final BooleanProperty supports;
+
+    @Override
+    public BooleanProperty isSupporting() {
+        return supports;
+    }
+}
