--- src/main/java/com/example/model/IOpinion.java
@@ -1,8 +1,0 @@
-package com.example.model;
-
-import javafx.beans.property.BooleanProperty;
-
-public interface IOpinion {
-    String getName();
-    BooleanProperty isSupporting();
-}
