--- src/main/java/com/example/model/IOpinions.java
@@ -1,6 +1,0 @@
-package com.example.model;
-
-public interface IOpinions {
-    void addOpinion(IOpinion opinion);
-    void removeOpinion(IOpinion opinion);
-}
