--- src/main/java/com/example/view/VertexSettingsView.java
@@ -1,4 +1,0 @@
-package com.example.view;
-
-public class VertexSettingsView {
-}
